Moved onMainThread() to ThreadSafety, removed IO scope
