+++ src/main/java/me/zeroeightsix/kami/util/threads/ThreadSafety.kt
@@ -4,6 +4,8 @@ import me.zeroeightsix.kami.event.ClientEvent
 import me.zeroeightsix.kami.event.ClientExecuteEvent
 import me.zeroeightsix.kami.event.SafeClientEvent
 import me.zeroeightsix.kami.event.SafeExecuteEvent
+import me.zeroeightsix.kami.util.Wrapper
+import java.util.concurrent.Callable
 
 fun ClientEvent.toSafe() =
     if (world != null && player != null && playerController != null && connection != null) SafeClientEvent(world, player, playerController, connection)
@@ -13,10 +15,35 @@ fun ClientExecuteEvent.toSafe() =
     if (world != null && player != null && playerController != null && connection != null) SafeExecuteEvent(world, player, playerController, connection, this)
     else null
 
-fun <R> runSafe(block: SafeClientEvent.() -> R) : R? {
+fun <R> runSafe(block: SafeClientEvent.() -> R): R? {
     return ClientEvent().toSafe()?.let { block(it) }
 }
 
-suspend fun <R> runSafeSuspend(block: suspend SafeClientEvent.() -> R) : R? {
+suspend fun <R> runSafeSuspend(block: suspend SafeClientEvent.() -> R): R? {
     return ClientEvent().toSafe()?.let { block(it) }
 }
+
+/**
+ * Run [block] on Minecraft main thread (Client thread) and wait for its result while blocking the current thread.
+ *
+ * @throws Exception if an exception thrown during [block] execution
+ *
+ * @see [onMainThreadSafe]
+ */
+fun <R> onMainThread(block: ClientEvent.() -> R) =
+    Wrapper.minecraft.addScheduledTask(Callable {
+        runCatching { ClientEvent().block() }
+    }).get().getOrThrow()
+
+/**
+ * Run [block] on Minecraft main thread (Client thread) and wait for its result while blocking the current thread.
+ * The [block] will the called with a [SafeClientEvent] to ensure null safety
+ *
+ * @throws Exception if an exception thrown during [block] execution
+ *
+ * @see [onMainThread]
+ */
+fun <R> onMainThreadSafe(block: SafeClientEvent.() -> R) =
+    Wrapper.minecraft.addScheduledTask(Callable {
+        runCatching { ClientEvent().toSafe()?.block() }
+    }).get().getOrThrow()
